+++ app/Http/routes.php
@@ -149,7 +149,7 @@ function()
             App::abort(404);
         }
 
-        $related_items = (!empty($item->related_work)) ? Item::where('related_work', '=', $item->related_work)->where('author', '=', $item->author)->whereNotNull('iipimg_url')->orderBy('related_work_order')->lists('iipimg_url')->toArray() : [];
+        $related_items = $item->getRelatedIIPImgUrls();
 
         return view('zoom', array('item' => $item, 'related_items' => $related_items));
     });
@@ -236,6 +236,24 @@ function()
 
         return view('dielo', array('item' => $item, 'more_items' => $more_items, 'previous' => $previous, 'next' => $next));
     });
+
+    Route::get('dielo/nahlad/{id}/{width}', function ($id, $width) {
+
+        if (
+            ($width <= 800) &&
+            Item::where('id', '=', $id)->exists()
+        ) {
+            // disable resizing when requesting 800px width
+            $width = ($width == 800) ? false : $width;
+            $resize_method = 'widen';
+            $imagePath = public_path() . Item::getImagePathForId($id, false, $width, $resize_method);
+
+            return response()->file($imagePath);
+        }
+
+        return App::abort(404);
+
+    })->where('width', '[0-9]+')->name('dielo.nahlad');
 
     Route::controller('patternlib', 'PatternlibController');