+++ public/js/webumenia.js
@@ -110,7 +110,7 @@ $(document).ready(function(){
     if (!is_touch_device) {
         $("[data-toggle='tooltip']").tooltip();
     }
-    
+
     $('#top a').click(function(){
         $('html,body').animate({scrollTop:0},'slow');return false;
     });
@@ -119,12 +119,12 @@ $(document).ready(function(){
     authors.initialize();
     articles.initialize();
     collections.initialize();
-    
+
     $('#search').typeahead(
     {
       hint: true,
       highlight: true,
-      minLength: 2      
+      minLength: 2
     },
     {
       name: 'authors',
@@ -179,7 +179,7 @@ $(document).ready(function(){
       }
     }).bind("typeahead:selected", function(obj, datum, name) {
         switch (name) {
-            case 'authors': 
+            case 'authors':
                 window.location.href = "/autor/" + datum.id;
                 break;
             case 'articles':
@@ -197,13 +197,10 @@ $(document).ready(function(){
 
 function spravGrid($container) {
     $container.isotope({
-        itemSelector : '.item',
-        masonry: {
-            isFitWidth: true,
-            gutter: 20
-        }
+        itemSelector: '.item',
+        layoutMode: 'masonry'
     });
-} 
+}
 
 var light_style = [
     {
@@ -315,3 +312,7 @@ var light_style = [
         ]
     }
 ];
+
+function isIE() { return navigator.userAgent.match(/Edge\/|Trident\/|MSIE /); }
+
+function isMobileSafari() { return navigator.userAgent.match(/(iPod|iPhone|iPad)/); }