+++ CHANGELOG.md
@@ -2,15 +2,39 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+### Changed
+- Favicons & Open Graph tags include components
+- positioned image reference strip on right side of zoom-viewer
+
+### Added
+- zoom-viewer component
+- Moravian gallery importer
+
+## [1.3.1] - 2018-03-09
+### Fixed
+- OAI-PMH harvester for translated attributes
+
+## [1.3.0] - 2018-02-20
 ### Added
 - Importer module
 - Czech language translations
 - Pattern Library at /patternlib
 - Tests setup
+- Option to run all importers
+- Route for requesting resized images at /dielo/nahlad/{id}/{width}
+- Lazy loading images in catalog and collection
+- Responsive image support in catalog and collection
+
+### Changed
+- Enabled reproduction orders with mounting / framing
+- Updated isotope-layout
+- Link to Public domain statement on CC website
 
 ### Fixed
 - Prevent upsize when uploading item image
 - Featured-article image container width bug
+- Two columns bug in iPhone6/SE
+- Artwork detail in iOS
 
 ## [1.2.9] - 2017-12-05
 ### Changed
@@ -38,6 +62,7 @@ All notable changes to this project will be documented in this file[^1].
 - LICENSE from Public Domain into MIT
 
 [Unreleased]: https://github.com/SlovakNationalGallery/web-umenia-2/compare/master...develop
+[1.3.0]: https://github.com/SlovakNationalGallery/web-umenia-2/pull/52
 [1.2.6]: https://github.com/SlovakNationalGallery/web-umenia-2/pull/31